Tests: Split utility methods into custom test case (#148)
* Tests: Split utility methods into custom test case

Moving the customer wrapper methods for the WP Unit Test factory methods into a test case, is the first step in being able to have multiple unit test classes, which can all extend from this new test case.

* Tests: Avoid nested filter callback

For "one-off" instances like this, which won't need to be updated by third-parties, the use of a closure is fine.

With the presence of a namespace in tests file, the previous callback reference was broken.
